Implement main bakesale.sh script, which should be sourced.
Source the `bakesale.sh` script from a deployment script, to run
bakesale. The main script displays things in a friendly manner, creates
the cakebox, and sets up `bakesale()`.

Call `bakesale()` with parameters to utilise the helper scripts, which
get included. The first two parameters, `$stage` and `$cmd`, are used to
determine where to look for the helper script -- nothing fancy.

Most parameters are expected to be passed through to stage helper
scripts; however, there are the following variables, which are
accessible anywhere:

- `$bakesale_script` :
  The path of the deployment script itself. I'm not sure this is all
  that useful.

- `$bakesale_source` :
  The path of the main `bakesale.sh` script, which is used for sourcing
  stage helper scripts.

- `$bakesale_cakebox` :
  The cakebox used for packing the current cake. This is where all
  preparation takes place, before the cakebox is delivered to bakesales
  using something like rsync+ssh.