Add bake stage helper script git.
This assists with packing the cakebox with a git repository. It is
expected to be the most commonly used case. The repository is cloned
locally, to save wrestling with permissions server-side (if indeed the
cakebox is being deployed to a server, which it might not be), cloning
all submodules, and wiping out the `.git/` directory. This is a little
upsetting, but works well with submodules.

e.g.

    bakesale bake git git://github.com/tiredpixel/tunefl.git