Add carry stage helper script rsyncssh, for simple rsync+ssh delivery.
This uses rsync+ssh for delivering the cakebox to the bakesale in a
simple manner. There is no serverly magic involving `current/`,
`releases/`, `shared/` directories, as indeed delivery to a server is
just one option. Perhaps counter-intuitively, the first string of
parameters is passed to ssh, and the second to rsync. This enables the
second set to be omitted.

e.g.

    bakesale carry rsyncssh tunefl@example.com:tunefl/

e.g.

    bakesale carry rsyncssh tunefl@example.com:tunefl/ "--exclude vendor/bundle/ --exclude .bundle/ --exclude bin/"