Add wave stage helper script ssh_foreman.
This uses ssh+foreman to ssh into a server, and use foreman, installing
if not already installed. `foreman export` is perhaps the most likely
use case. The script is similar to `ssh_rvm_foreman`, except that it is
designed for systems not running RVM (such as that don't run Ruby for
any application itself).

e.g.

    bakesale wave ssh_foreman \
        peace@example.com \
        "export --root octopus/ --app octopus --user octopus upstart /etc/init"